Add test script
This is a simple test script I wrote to attempt to verify what I was told in my support ticket. What I was told was wrong.

It creates a simple server that listens for a GitHub webhook push, alters it with the addition of username and avatar, and then passes it along to the originally intended webhook endpoint.

This clearly demonstrates that the use of that butt-ugly icon is a hardcoded override present on Discord's side of the equation.